Edit product page tests

+++ database/migrations/2015_09_13_083656_create_bill_application_products_table.php
@@ -21,7 +21,7 @@ class CreateBillApplicationProductsTable extends Migration {
             $table->bigIncrements('id');
             $table->bigInteger('bill_id')->unsigned();
             $table->bigInteger('product_id')->unsigned();
-            $table->tinyInteger('page')->default(4);
+            $table->smallInteger('page')->default(4);
             $table->tinyInteger('quantity');
             $table->float('price');
             $table->tinyInteger('discount');
